+++ src/OpenTelemetry.Instrumentation.Owin/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Skip unnecessary OWIN processing when tracing is not enabled.
+  ([#4919](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4919))
+
 ## 1.17.0-beta.1
 
 Released 2026-Jul-17